Set up SSH public key authentication
Using SSH public key authentication to connect to a remote system is a robust, more secure alternative to logging in with an account password or passphrase. 
SSH public key authentication relies on asymmetric cryptographic algorithms that generate a pair of separate keys (a key pair), one "private" and the other "public". 
You keep the private key a secret and store it on the computer you use to connect to the remote system. 
Conceivably, you can share the public key with anyone without compromising the private key; you store it on the remote system in a .ssh/authorized_keys directory.